Store corrected/raw image data as float instead of double
DIALS integration is ultimately float precision, so the image-sized
arrays returned by the ImageSet are held in double at twice the memory
they need. Convert get_corrected_data() and the persistent raw-data
cache from Image<double> to Image<float> to roughly halve the resident
size of these per-image buffers (the change that produced the observed
~10% RSS drop in dials.stills_process).

Precision is preserved where it matters: the gain and pedestal
corrections are still read from their double lookup tables and the
subtraction/division is done in double, with only the final corrected
pixel stored as float. nxmx_writer now accepts flex.float panel data in
addition to flex.double so float images can still be written out.

This is the foundational change; the dials repo is updated in lock-step
to consume the float get_corrected_data() and to down-precision the
image-sized intermediates built from it.

+++ src/dxtbx/boost_python/imageset_ext.cc
@@ -451,7 +451,7 @@ namespace dxtbx { namespace boost_python {
   }
 
   boost::python::tuple ImageSet_get_corrected_data(ImageSet &self, std::size_t index) {
-    return image_as_tuple<double>(self.get_corrected_data(index));
+    return image_as_tuple<float>(self.get_corrected_data(index));
   }
 
   boost::python::tuple ImageSet_get_gain(ImageSet &self, std::size_t index) {
